修改烘干塔接货人接口

+++ src/main/java/com/demai/cornel/vo/task/GetOrderListReq.java
@@ -9,6 +9,6 @@ import lombok.Data;
 @Data
 public class GetOrderListReq {
     private String orderId;
-    private Integer orderTyp;//获取带收粮  已取消订单    订单状态详情 见wiki：https://yapi.demeteria.com/project/13/wiki
+    private Long orderTyp;//获取带收粮  已取消订单    订单状态详情 见wiki：https://yapi.demeteria.com/project/13/wiki
     private Integer pgSize;
 }
